# webapp/plugins/notificationstore/api/swagger/v2/swagger.json
{"swagger":"2.0","info":{"description":"This API can be used by client application to manage Notifications","version":"V2","title":"Notification store API"},"tags":[{"name":"restnotificationstorev3category"},{"name":"restnotificationstorev3demand"},{"name":"restnotificationstorev3demandType"},{"name":"restnotificationstorev3"}],"consumes":["application/json"],"produces":["application/json"],"paths":{"/rest/notificationstore/v3/category/list":{"get":{"tags":["restnotificationstorev3category"],"summary":"Get the list of categories","operationId":"getDemandCategoryList","produces":["application/json"],"responses":{"200":{"description":"Success"},"204":{"description":"No content"},"400":{"description":"Bad request or missing mandatory parameters"},"403":{"description":"Failure"}}}},"/rest/notificationstore/v3/category":{"post":{"tags":["restnotificationstorev3category"],"operationId":"createDemandCategory","produces":["application/json"],"parameters":[{"name":"code","in":"formData","required":false,"type":"string"},{"name":"label","in":"formData","required":false,"type":"string"},{"name":"isDefault","in":"formData","required":false,"type":"string"}],"responses":{"default":{"description":"successful operation"}}}},"/rest/notificationstore/v3/category/{id}":{"get":{"tags":["restnotificationstorev3category"],"summary":"Get a category","operationId":"getDemandCategory","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Category id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"},"404":{"description":"Not found"}}},"put":{"tags":["restnotificationstorev3category"],"operationId":"modifyDemandCategory","produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"},{"name":"code","in":"formData","required":false,"type":"string"},{"name":"label","in":"formData","required":false,"type":"string"},{"name":"isDefault","in":"formData","required":false,"type":"string"}],"responses":{"default":{"description":"successful operation"}}},"delete":{"tags":["restnotificationstorev3category"],"operationId":"deleteDemandCategory","produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"default":{"description":"successful operation"}}}},"/rest/notificationstore/v3/demand/list":{"get":{"tags":["restnotificationstorev3demand"],"summary":"Get demand list for a customer Id","operationId":"getListDemand","produces":["application/json"],"parameters":[{"name":"idDemandType","in":"query","description":"Demand type id","required":false,"type":"string"},{"name":"index","in":"query","description":"Index of pagination","required":false,"type":"string"},{"name":"limitResult","in":"query","description":"Max of responses requested","required":false,"type":"string"},{"name":"customerId","in":"query","description":"Id of the customer","required":false,"type":"string"},{"name":"notificationType","in":"query","description":"Type of notification","required":false,"type":"string"}],"responses":{"200":{"description":"Success"},"400":{"description":"Bad request or missing mandatory parameters"},"403":{"description":"Failure"}}}},"/rest/notificationstore/v3/demand/status":{"get":{"tags":["restnotificationstorev3demand"],"summary":"Get demand list for a customer Id by status","operationId":"getListOfDemandByStatus","produces":["application/json"],"parameters":[{"name":"idDemandType","in":"query","description":"Demand type id","required":false,"type":"string"},{"name":"index","in":"query","description":"Index of pagination","required":false,"type":"string"},{"name":"limitResult","in":"query","description":"Max of responses requested","required":false,"type":"string"},{"name":"customerId","in":"query","description":"Id of the customer","required":false,"type":"string"},{"name":"listStatus","in":"query","description":"List of requested status","required":false,"type":"string"},{"name":"notificationType","in":"query","description":"Type of notification","required":false,"type":"string"},{"name":"categoryCode","in":"query","description":"List of requested category codes","required":false,"type":"string"}],"responses":{"200":{"description":"Success (with or without result)"},"400":{"description":"Bad request or missing mandatory parameters"},"403":{"description":"Failure"}}}},"/rest/notificationstore/v3/demand/{customerId}":{"delete":{"tags":["restnotificationstorev3demand"],"operationId":"doDeleteAllDemands","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"customerId","in":"path","required":true,"type":"string"}],"responses":{"default":{"description":"successful operation"}}}},"/rest/notificationstore/v3/demandType":{"get":{"tags":["restnotificationstorev3demandType"],"summary":"Get demand type list (stable, must be cached)","operationId":"getDemandTypeList","produces":["application/json"],"parameters":[{"name":"direct","in":"query","description":"(Deprecated)","required":false,"type":"string"}],"responses":{"200":{"description":"Success"},"400":{"description":"Bad request or missing mandatory parameters"},"403":{"description":"Failure"}}},"post":{"tags":["restnotificationstorev3demandType"],"operationId":"createDemandType","produces":["application/json"],"parameters":[{"name":"id_demand_type","in":"formData","required":false,"type":"string"},{"name":"label","in":"formData","required":false,"type":"string"},{"name":"url","in":"formData","required":false,"type":"string"},{"name":"app_code","in":"formData","required":false,"type":"string"},{"name":"category","in":"formData","required":false,"type":"string"}],"responses":{"default":{"description":"successful operation"}}}},"/rest/notificationstore/v3/demandType/{id}":{"get":{"tags":["restnotificationstorev3demandType"],"summary":"Get demand type (stable, must be cached)","operationId":"getDemandType","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Demand type id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"},"400":{"description":"Bad request or missing mandatory parameters"},"404":{"description":"Not found"},"403":{"description":"Failure"}}},"put":{"tags":["restnotificationstorev3demandType"],"operationId":"modifyDemandType","produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"},{"name":"id_demand_type","in":"formData","required":false,"type":"string"},{"name":"label","in":"formData","required":false,"type":"string"},{"name":"url","in":"formData","required":false,"type":"string"},{"name":"app_code","in":"formData","required":false,"type":"string"},{"name":"category","in":"formData","required":false,"type":"string"}],"responses":{"default":{"description":"successful operation"}}},"delete":{"tags":["restnotificationstorev3demandType"],"operationId":"deleteDemandType","produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"default":{"description":"successful operation"}}}},"/rest/notificationstore/v3/notification":{"post":{"tags":["restnotificationstorev3"],"operationId":"notification","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":false,"schema":{"type":"string"}}],"responses":{"default":{"description":"successful operation"}}}},"/rest/notificationstore/v3/notificationnotificationType":{"get":{"tags":["restnotificationstorev3"],"operationId":"getNotificationTypes","produces":["application/json"],"responses":{"default":{"description":"successful operation"}}}},"/rest/notificationstore/v3/notificationEvent":{"post":{"tags":["restnotificationstorev3"],"operationId":"notificationEvent","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":false,"schema":{"type":"string"}}],"responses":{"default":{"description":"successful operation"}}}},"/rest/notificationstore/v3/notification/list":{"get":{"tags":["restnotificationstorev3"],"summary":"Get the notifications of a demand","operationId":"getListNotification","produces":["application/json"],"parameters":[{"name":"idDemand","in":"query","description":"Demand id","required":false,"type":"string"},{"name":"idDemandType","in":"query","description":"Demand type id","required":false,"type":"string"},{"name":"customerId","in":"query","description":"Id of the customer","required":false,"type":"string"},{"name":"notificationType","in":"query","description":"Type of notification","required":false,"type":"string"}],"responses":{"200":{"description":"Success"},"400":{"description":"Bad request or missing mandatory parameters"},"403":{"description":"Failure"}}}}},"definitions":{"DemandDisplay":{"type":"object","properties":{"status":{"type":"string"},"demand":{"$ref":"#/definitions/demand"}}},"DemandResult":{"type":"object","properties":{"listDemandDisplay":{"type":"array","items":{"$ref":"#/definitions/DemandDisplay"}},"index":{"type":"string"},"paginator":{"type":"string"},"numberResult":{"type":"integer","format":"int32"},"status":{"$ref":"#/definitions/ResponseStatus"}}},"ResponseStatus":{"type":"object","properties":{"http_code":{"type":"integer","format":"int32","readOnly":true},"status":{"type":"string","readOnly":true,"enum":["OK","SUCCESS","INCOMPLETE_SUCCESS","BAD_REQUEST","UNAUTHORIZED","FAILURE","NOT_FOUND","CONFLICT","INTERNAL_SERVER_ERROR"]},"message":{"type":"string"},"message_key":{"type":"string"}}},"customer":{"type":"object","properties":{"attributeNames":{"type":"array","items":{"type":"string"}},"firstname":{"type":"string"},"lastname":{"type":"string"},"familyname":{"type":"string"},"fixedPhoneNumber":{"type":"string"},"mobilePhone":{"type":"string"},"birthDate":{"type":"string"},"id":{"type":"string"},"email":{"type":"string"},"connection_id":{"type":"string"},"customer_id":{"type":"string"}}},"demand":{"type":"object","properties":{"id":{"type":"string"},"type_id":{"type":"string"},"subtype_id":{"type":"string"},"reference":{"type":"string"},"status_id":{"type":"integer","format":"int32"},"max_step":{"type":"integer","format":"int32"},"current_step":{"type":"integer","format":"int32"},"customer":{"$ref":"#/definitions/customer"},"uid":{"type":"integer","format":"int32"},"modify_date":{"type":"integer","format":"int64"}}},"NotificationResult":{"type":"object","properties":{"notifications":{"type":"array","items":{"$ref":"#/definitions/notification"}},"index":{"type":"string"},"paginator":{"type":"string"},"numberResult":{"type":"integer","format":"int32"},"status":{"$ref":"#/definitions/ResponseStatus"}}},"backoffice_logging":{"type":"object","properties":{"message":{"type":"string"},"status_text":{"type":"string"}}},"broadcast_email":{"type":"object","properties":{"message":{"type":"string"},"bcc":{"type":"array","items":{"$ref":"#/definitions/email_address"}},"subject":{"type":"string"},"sender_email":{"type":"string"},"sender_name":{"type":"string"},"recipient":{"type":"array","items":{"$ref":"#/definitions/email_address"}},"cc":{"type":"array","items":{"$ref":"#/definitions/email_address"}}}},"email_address":{"type":"object","properties":{"address":{"type":"string"}}},"mydashboard":{"type":"object","properties":{"status_id":{"type":"integer","format":"int32"},"status_text":{"type":"string"},"message":{"type":"string"},"subject":{"type":"string"},"data":{"type":"string"},"sender_name":{"type":"string"}}},"notification":{"type":"object","properties":{"date":{"type":"integer","format":"int64"},"demand":{"$ref":"#/definitions/demand"},"backoffice":{"$ref":"#/definitions/backoffice_logging"},"customer_email":{"$ref":"#/definitions/user_email"},"sms":{"$ref":"#/definitions/sms"},"mydashboard":{"$ref":"#/definitions/mydashboard"},"broadcast_email":{"type":"array","items":{"$ref":"#/definitions/broadcast_email"}}}},"sms":{"type":"object","properties":{"message":{"type":"string"},"phone_number":{"type":"string"},"sender_name":{"type":"string"}}},"user_email":{"type":"object","properties":{"message":{"type":"string"},"bcc":{"type":"string"},"subject":{"type":"string"},"sender_email":{"type":"string"},"sender_name":{"type":"string"},"recipient":{"type":"string"},"cc":{"type":"string"}}}},"externalDocs":{"description":"Lutèce WIKI","url":"https://lutece.paris.fr/support/jsp/site/Portal.jsp?page=wiki"}}